+++ src/main/java/com/aetherteam/aether/integration/jei/categories/block/AbstractAetherBlockStateRecipeCategory.java
@@ -1,7 +1,7 @@
 package com.aetherteam.aether.integration.jei.categories.block;
 
 import com.aetherteam.aether.integration.jei.categories.block.AbstractBlockStateRecipeCategory;
-import com.aetherteam.nitrogen.recipe.recipes.AbstractBlockStateRecipe;
+import com.aetherteam.aether.recipe.blockstate.recipes.AbstractBlockStateRecipe;
 import mezz.jei.api.gui.drawable.IDrawable;
 import mezz.jei.api.helpers.IPlatformFluidHelper;
 import mezz.jei.api.recipe.types.IRecipeType;