+++ src/main/java/com/aetherteam/aether/integration/jei/categories/block/AbstractBlockStateRecipeCategory.java
@@ -1,9 +1,9 @@
 package com.aetherteam.aether.integration.jei.categories.block;
 
 import com.aetherteam.aether.integration.jei.categories.AbstractRecipeCategory;
-import com.aetherteam.nitrogen.recipe.BlockPropertyPair;
-import com.aetherteam.nitrogen.recipe.BlockStateIngredient;
-import com.aetherteam.nitrogen.recipe.recipes.AbstractBlockStateRecipe;
+import com.aetherteam.aether.recipe.blockstate.BlockPropertyPair;
+import com.aetherteam.aether.recipe.blockstate.BlockStateIngredient;
+import com.aetherteam.aether.recipe.blockstate.recipes.AbstractBlockStateRecipe;
 import mezz.jei.api.gui.builder.IRecipeLayoutBuilder;
 import mezz.jei.api.gui.builder.ITooltipBuilder;
 import mezz.jei.api.gui.drawable.IDrawable;