+++ src/main/java/com/aetherteam/aether/recipe/blockstate/BlockStateIngredient.java
@@ -1,4 +1,4 @@
-package com.aetherteam.nitrogen.recipe;
+package com.aetherteam.aether.recipe.blockstate;
 
 import com.mojang.datafixers.util.Either;
 import com.mojang.serialization.Codec;